core 모듈 context는core가 관리하도록 presentation module은 import해서 쓴다
